src/requirements.txt
python-telegram-bot[job-queue]
requests
Pillow
python-dotenv
#apscheduler
pytz
sqlalchemy
pydantic
beautifulsoup4
bs4
lxml